producto falta eliminar

+++ models/client.go
@@ -1,7 +1,7 @@
 package models
 
 type Client struct {
-	ID       int `json:"_id"`
-	IdPerson int `json:"idPerson"`
+	ID       int    `json:"_id"`
+	IdPerson int64  `json:"idPerson"`
 	Type     string `json:"type"`
 }
